branch:      2.0 backported the following to 2.0: * EIGEN_ALIGN and EIGEN_DONT_ALIGN and the corresponding logic in Macros.h   (instead of using EIGEN_ARCH_WANTS_ALIGNMENT) * The body of ei_aligned_malloc and ei_aligned_free
The reason for this backporting is that a user complained that with eigen 2.0 he got a warning at Memory.h:81 that the return value of posix_memalign was not used, and that function was declared with an attribute warn_unused_result.

Looking at this, it seemed that the body of this function was already overly complicated, and fixing this warning made it even worse, while the devel branch had a much simpler body and didn't suffer from that problem.

Then it was necessary to update ei_aligned_free too, and to backport EIGEN_ALIGN.

Inch' Allah....